## What does this PR do?
Adds benchmarks for trace load and parsing  #48 

## Before submitting

- [x] Was this discussed/approved via a Github issue? (no need for typos, doc improvements)
  - [ ] N/A
- [ ] Did you write any new necessary tests?
  - [x] N/A
- [ ] Did you make sure to update the docs?
  - [x] N/A
- [ ] Did you update the [changelog](https://github.com/facebookresearch/HolisticTraceAnalysis/blob/main/CHANGELOG.md)?
  - [ ] N/A 
    not sure?

## Benchmark results

Run using 
```
python3 benchmarks/trace_load_benchmark.py -p 10 -l 1
```

Before  Anupam's PR #48  optimizing out the get_memory_size we see the load/parsing times were
```
parse[tests/data/vision_transformer]: Mean +- std dev: 12.4 sec +- 0.5 sec
parse[tests/data/inference_single_rank]: Mean +- std dev: 15.0 sec +- 0.5 sec
```

After his PR we now see the time reduced for loading
```
parse[tests/data/vision_transformer]: Mean +- std dev: 6.24 sec +- 0.19 sec
parse[tests/data/inference_single_rank]: Mean +- std dev: 9.25 sec +- 0.19 sec
```